% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompressedSensing_solver.R
\name{CompressedSensing_solver}
\alias{CompressedSensing_solver}
\title{Solve a Compressed Sensing Problem Using the Parametric Simplex Method}
\usage{
CompressedSensing_solver(X, y, max_it = 50, lambda_threshold = 0.01)
}
\arguments{
\item{X}{An \code{n} by \code{d} data matrix.}

\item{y}{A length \code{n} response vector.}

\item{max_it}{Maximum number of iterations for the solution path. The default is \code{50}.}

\item{lambda_threshold}{The algorithm stops when the regularization parameter falls below this threshold. The default value is \code{0.01}.}
}
\description{
Solve a Compressed Sensing Problem Using the Parametric Simplex Method
}
\note{
The returned value is similar to that of \code{\link{Dantzig_solver}}.
}
\seealso{
\code{\link{primal-package}}, \code{\link{Dantzig_solver}}
}
