% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dantzig_solver.R
\name{Dantzig_solver}
\alias{Dantzig_solver}
\title{Solve a Dantzig Selector Problem Using the Parametric Simplex Method}
\usage{
Dantzig_solver(X, y, max_it = 50, lambda_threshold = 0.01)
}
\arguments{
\item{X}{An \code{n} by \code{d} data matrix.}

\item{y}{A length \code{n} response vector.}

\item{max_it}{Maximum number of iterations for the solution path. The default is \code{50}.}

\item{lambda_threshold}{The algorithm stops when the regularization parameter falls below this threshold. The default value is \code{0.01}.}
}
\value{
An object with S3 class \code{"primal"} is returned:
\item{data}{
  The \code{n} by \code{d} data matrix from the input.
}
\item{response}{
  The length \code{n} response vector from the input.
}
\item{beta}{
  A matrix of coefficient estimates whose columns correspond to regularization parameters along the solution path.
}
\item{df}{
  The degrees of freedom (number of nonzero coefficients) along the solution path.
}
\item{value}{
  The sequence of optimal objective function values corresponding to each \code{lambda}.
}
\item{iterN}{
  The number of iterations performed.
}
\item{lambda}{
  The sequence of regularization parameters obtained along the solution path.
}
\item{type}{
  The type of the problem, e.g., \code{"Dantzig"} or \code{"SparseSVM"}.
}
}
\description{
Solve a Dantzig Selector Problem Using the Parametric Simplex Method
}
\seealso{
\code{\link{primal-package}}
}
