% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.primal.R
\name{plot.primal}
\alias{plot.primal}
\title{Plot a "primal" Object}
\usage{
\method{plot}{primal}(x, n = NULL, ...)
}
\arguments{
\item{x}{An object with S3 class \code{"primal"}.}

\item{n}{If \code{NULL}, all three plots are shown together. If \code{n} is a number (1, 2, or 3), only the corresponding plot is shown.}

\item{\dots}{Additional arguments (currently unused).}
}
\description{
Plot the regularization path and parameters obtained from the parametric simplex method.
}
\seealso{
\code{\link{Dantzig_solver}}, \code{\link{SparseSVM_solver}}
}
