% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.primal.R
\name{print.primal}
\alias{print.primal}
\title{Print a "primal" Object}
\usage{
\method{print}{primal}(x, ...)
}
\arguments{
\item{x}{An object with S3 class \code{"primal"}.}

\item{\dots}{Additional arguments (currently unused).}
}
\description{
Print summary information including the model type, regularization parameter path, and degrees of freedom along the solution path.
}
\seealso{
\code{\link{Dantzig_solver}}, \code{\link{SparseSVM_solver}}
}
