% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantileRegression_solver.R
\name{QuantileRegression_solver}
\alias{QuantileRegression_solver}
\title{Solve a Quantile Regression Problem Using the Parametric Simplex Method}
\usage{
QuantileRegression_solver(X, y, max_it = 50, lambda_threshold = 0.01,
  tau = 0.5)
}
\arguments{
\item{X}{An \code{n} by \code{d} data matrix.}

\item{y}{A length \code{n} response vector.}

\item{max_it}{Maximum number of iterations for the solution path. The default is \code{50}.}

\item{lambda_threshold}{The algorithm stops when the regularization parameter falls below this threshold. The default value is \code{0.01}.}

\item{tau}{The quantile level for the regression. The default is \code{0.5}.}
}
\description{
Solve a Quantile Regression Problem Using the Parametric Simplex Method
}
\note{
The returned value is similar to that of \code{\link{Dantzig_solver}}.
}
\seealso{
\code{\link{primal-package}}, \code{\link{Dantzig_solver}}
}
