% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.primal.R
\name{coef.primal}
\alias{coef.primal}
\title{Coef function for S3 class "primal"}
\usage{
\method{coef}{primal}(str, n = NULL)
}
\arguments{
\item{str}{An object with S3 class \code{"primal"}.}

\item{n}{The index of the wanted parameter.}
}
\description{
Print the estimated solution correspond to a specific parameter.
}
\seealso{
\code{\link{Dantzig_solver}}, \code{\link{SparseSVM_solver}}
}
