% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompressedSensing_solver.R
\name{CompressedSensing_solver}
\alias{CompressedSensing_solver}
\title{Solve given compressed sensing problem in parametric simplex method}
\usage{
CompressedSensing_solver(X, y, max_it = 50, lambda_threshold = 0.01)
}
\arguments{
\item{X}{\code{x} is an \code{n} by \code{d} data matrix}

\item{y}{\code{y} is a length \code{n} response vector}

\item{max_it}{This is the number of the maximum path length one would like to achieve. The default length is \code{50}.}

\item{lambda_threshold}{The parametric simplex method will stop when the calculated parameter is smaller than lambda. The default value is \code{0.01}.}
}
\description{
Solve given compressed sensing problem in parametric simplex method
}
\note{
The returned value is similar with \code{Dantzig_solver()}
}
\seealso{
\code{\link{primal-package}}, \code{\link{Dantzig_solver}}
}
