% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.primal.R
\name{plot.primal}
\alias{plot.primal}
\title{Plot function for S3 class "primal"}
\usage{
\method{plot}{primal}(str, n = NULL, ...)
}
\arguments{
\item{str}{An object with S3 class \code{"primal"}}

\item{n}{If \code{n = NULL}, three graph will be shown together. If \code{n} is a number, then the corresponding graph will be shown.}

\item{\dots}{System reserved (No specific usage)}
}
\description{
Plot regularization path and parameter obtained from the algorithm.
}
\seealso{
\code{\link{Dantzig_solver}}, \code{\link{SparseSVM_solver}}
}
