% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMAL-package.R
\docType{package}
\name{PRIMAL-package}
\alias{PRIMAL-package}
\alias{_PACKAGE}
\alias{primal-package}
\title{Parametric Simplex Method for Sparse Learning}
\description{
A package for parametric simplex method for sparse learning
}
\details{
\tabular{ll}{
  Package: \tab PRIMAL\cr
  Type: \tab Package\cr
  Version: \tab 1.0.0\cr
  Date: \tab 2019-08-15\cr
}
The package "PRIMAL" provides 4 main functions:\cr
(1) The dantzig selector solver applying simplex method. Please refer to \code{\link{Dantzig_solver}}.\cr
(2) The sparse SVM solver applying simplex method. Please refer to \code{\link{SparseSVM_solver}}.\cr
(3) The compressed sensing solver. Please refer to \code{\link{CompressedSensing_solver}}.\cr
(4) The quantile regression solver. Please refer to \code{\link{QuantileRegression_solver}}.\cr
}
\seealso{
\code{\link{plot.primal}}, \code{\link{print.primal}}, \code{\link{coef.primal}}
}
\author{
Qianli Shen, Zichong Li \cr
}
