% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.primal.R
\name{print.primal}
\alias{print.primal}
\title{Print function for S3 class "primal"}
\usage{
\method{print}{primal}(str, ...)
}
\arguments{
\item{str}{An object with S3 class \code{"primal"}.}

\item{\dots}{System reserved (No specific usage)}
}
\description{
Print the information about the model usage, the parameter path, degree of freedom of the solution path.
}
\seealso{
\code{\link{Dantzig_solver}}, \code{\link{SparseSVM_solver}}
}
